+++ packages/avalanche-module/src/handlers/get-network-fee/get-network-fee.ts
@@ -6,17 +6,15 @@ import { getProvider } from '../../utils/get-provider';
  */
 export async function getNetworkFee({
   isTestnet,
-  isDevnet,
   vmName,
 }: {
   isTestnet: boolean;
-  isDevnet?: boolean; // TODO(@meeh0w): remove `isDevnet` after E-upgrade activation on Fuji
   vmName: NetworkVMType;
 }): Promise<NetworkFees> {
-  const provider = await getProvider({ isTestnet, isDevnet });
+  const provider = await getProvider({ isTestnet });
 
-  // Return static fees for X-Chain and pre-Etna P-Chain
-  if (vmName === NetworkVMType.AVM || !provider.isEtnaEnabled()) {
+  // Return static fees for X-Chain
+  if (vmName === NetworkVMType.AVM) {
     // this is 0.001 Avax denominated in nAvax, taken from https://docs.avax.network/reference/standards/guides/txn-fees#fee-schedule
     return {
       baseFee: BigInt(1000000),