+++ packages/evm-module/src/handlers/eth-send-transaction/eth-send-transaction.test.ts
@@ -95,6 +95,7 @@ const testParams = {
   nonce: '12',
   gas: '0x5208',
   chainId: '0x1',
+  accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
 };
 const testDapp = { url: 'https://example.com', name: 'dapp', icon: 'icon' };
 const testRequestParams = () => ({
@@ -164,6 +165,7 @@ const signingData = {
     data: '0xdata',
     value: '0xvalue',
     chainId: 1,
+    accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
   },
 };
 
@@ -204,7 +206,16 @@ describe('eth_sendTransaction handler', () => {
     });
     mockParseRequestParams.mockReturnValue({
       success: true,
-      data: [{ from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue', nonce: '12' }],
+      data: [
+        {
+          from: '0xfrom',
+          to: '0xto',
+          data: '0xdata',
+          value: '0xvalue',
+          nonce: '12',
+          accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+        },
+      ],
     });
     mockEstimateGasLimit.mockResolvedValue(21000);
 
@@ -222,7 +233,13 @@ describe('eth_sendTransaction handler', () => {
 
     expect(mockEstimateGasLimit).toHaveBeenCalledWith({
       provider: mockProvider,
-      transactionParams: { from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue' },
+      transactionParams: {
+        from: '0xfrom',
+        to: '0xto',
+        data: '0xdata',
+        value: '0xvalue',
+        accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+      },
     });
 
     expect(mockApprovalController.requestApproval).toHaveBeenCalledWith({
@@ -244,7 +261,16 @@ describe('eth_sendTransaction handler', () => {
     });
     mockParseRequestParams.mockReturnValue({
       success: true,
-      data: [{ from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue', gas: '0x5208' }],
+      data: [
+        {
+          from: '0xfrom',
+          to: '0xto',
+          data: '0xdata',
+          value: '0xvalue',
+          gas: '0x5208',
+          accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+        },
+      ],
     });
     mockGetNonce.mockResolvedValue(12);
 
@@ -283,7 +309,15 @@ describe('eth_sendTransaction handler', () => {
     });
     mockParseRequestParams.mockReturnValue({
       success: true,
-      data: [{ from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue' }],
+      data: [
+        {
+          from: '0xfrom',
+          to: '0xto',
+          data: '0xdata',
+          value: '0xvalue',
+          accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+        },
+      ],
     });
     mockGetNonce.mockResolvedValue(12);
     mockEstimateGasLimit.mockResolvedValue(21000);
@@ -306,7 +340,13 @@ describe('eth_sendTransaction handler', () => {
 
     expect(mockEstimateGasLimit).toHaveBeenCalledWith({
       provider: mockProvider,
-      transactionParams: { from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue' },
+      transactionParams: {
+        from: '0xfrom',
+        to: '0xto',
+        data: '0xdata',
+        value: '0xvalue',
+        accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+      },
     });
 
     expect(mockApprovalController.requestApproval).toHaveBeenCalledWith({
@@ -342,8 +382,10 @@ describe('eth_sendTransaction handler', () => {
             simulation: {
               status: 'Success',
               account_summary: {
-                exposures: [
+                traces: [
                   {
+                    trace_type: 'ExposureTrace',
+                    type: 'ERC20ExposureTrace',
                     asset: {
                       type: TokenType.ERC20,
                       address: '0xTokenAddress',
@@ -352,10 +394,10 @@ describe('eth_sendTransaction handler', () => {
                       decimals: 18,
                       logo_url: 'logo_url',
                     },
-                    spenders: {
-                      '0xSpenderAddress': {
-                        exposure: [{ raw_value: '1', usd_price: '1' }],
-                      },
+                    spender: '0xSpenderAddress',
+                    exposed: {
+                      raw_value: '1',
+                      usd_price: 1,
                     },
                   },
                 ],
@@ -382,7 +424,17 @@ describe('eth_sendTransaction handler', () => {
 
     mockParseRequestParams.mockReturnValue({
       success: true,
-      data: [{ from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue', nonce: '12', gas: '0x5208' }],
+      data: [
+        {
+          from: '0xfrom',
+          to: '0xto',
+          data: '0xdata',
+          value: '0xvalue',
+          nonce: '12',
+          gas: '0x5208',
+          accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+        },
+      ],
     });
 
     const updateTx = jest.fn();
@@ -602,7 +654,17 @@ const testWithValidationResultType = async (resultType: 'Warning' | 'Error' | 'M
 
   mockParseRequestParams.mockReturnValue({
     success: true,
-    data: [{ from: '0xfrom', to: '0xto', data: '0xdata', value: '0xvalue', nonce: '12', gas: '0x5208' }],
+    data: [
+      {
+        from: '0xfrom',
+        to: '0xto',
+        data: '0xdata',
+        value: '0xvalue',
+        nonce: '12',
+        gas: '0x5208',
+        accessList: [{ address: '0x123', storageKeys: ['0xkey1', '0xkey2'] }],
+      },
+    ],
   });
 
   const requestParams = testRequestParams();