+++ packages/hvm-module/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @avalabs/hvm-module
 
+## 1.4.6
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.4.6
+
+## 1.4.5
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.4.5
+
+## 1.4.4
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.4.4
+
+## 1.4.3
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.4.3
+
+## 1.4.2
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.4.2
+
+## 1.4.1
+
+### Patch Changes
+
+- 0a75fcc: upgrade @avalabs/core-utils-sdk in hvm module
+- Updated dependencies [0a75fcc]
+- Updated dependencies [0a75fcc]
+  - @avalabs/vm-module-types@1.4.1
+
+## 1.4.0
+
+### Minor Changes
+
+- 8969b33: Add EIP-2930 support
+
+### Patch Changes
+
+- 0954386: Fix broken lock file
+- Updated dependencies [0954386]
+- Updated dependencies [8969b33]
+  - @avalabs/vm-module-types@1.4.0
+
+## 1.3.0
+
+### Patch Changes
+
+- @avalabs/vm-module-types@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes